# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListStudentsResponse:
    properties:
      students:
        items:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.StudentResponse'
        type: array
      total:
        type: integer
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.RegisterStudentRequest:
    properties:
      first_name:
        type: string
      grade:
        type: string
      last_name:
        type: string
      pickup_address:
        type: string
      pickup_location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
      school_id:
        type: string
    required:
    - first_name
    - last_name
    - school_id
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.StudentResponse:
    properties:
      first_name:
        type: string
      grade:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      pickup_address:
        type: string
      pickup_location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
      school_id:
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.UpdateStudentRequest:
    properties:
      first_name:
        type: string
      grade:
        type: string
      last_name:
        type: string
      pickup_address:
        type: string
      pickup_location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for the gateway service of the School Tracking System.
  title: Gateway Service API
  version: "1.0"
paths:
...
  /fleet/students:
    get:
      description: Retrieve a paginated list of students
      parameters:
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListStudentsResponse'
      security:
      - BearerAuth: []
      summary: List students
      tags:
      - fleet
    post:
      consumes:
      - application/json
      description: Create a new student record
      parameters:
      - description: Register Student Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.RegisterStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.StudentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a student
      tags:
      - fleet
  /fleet/students/{id}:
    delete:
      description: Mark a student as inactive
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a student
      tags:
      - fleet
    get:
      description: Retrieve student details by UUID
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.StudentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a student by ID
      tags:
      - fleet
    put:
      consumes:
      - application/json
      description: Modify student details
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update Student Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.StudentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a student
      tags:
      - fleet
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"